requirements.txt
fastapi
uvicorn
requests
dotenv
google-auth
google-auth-oauthlib
google-api-python-client
sqlalchemy
mysqlclient
jose